requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
# created-by: conda 24.9.2
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=2.1.0=pypi_0
accelerate=1.0.1=pypi_0
addict=2.4.0=pypi_0
aiofiles=23.2.1=pypi_0
aiohttp=3.9.5=pypi_0
aiosignal=1.3.1=pypi_0
albumentations=1.3.0=pypi_0
altair=5.3.0=pypi_0
annotated-types=0.7.0=pypi_0
antlr4-python3-runtime=4.8=pypi_0
anyio=4.4.0=pypi_0
asttokens=3.0.0=pypi_0
async-timeout=4.0.3=pypi_0
attrs=23.2.0=pypi_0
backcall=0.2.0=pypi_0
backports-zoneinfo=0.2.1=pypi_0
basicsr=1.4.2=pypi_0
blas=1.0=mkl
blinker=1.8.2=pypi_0
braceexpand=0.1.7=pypi_0
brotli-python=1.0.9=py38h6a678d5_8
bzip2=1.0.8=h5eee18b_6
ca-certificates=2024.3.11=h06a4308_0
cachetools=5.3.3=pypi_0
certifi=2024.6.2=py38h06a4308_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.1.7=pypi_0
comm=0.2.2=pypi_0
configargparse=1.7=pypi_0
contourpy=1.1.1=pypi_0
cudatoolkit=11.3.1=h2bc3f7f_2
cycler=0.12.1=pypi_0
dash=2.18.2=pypi_0
dash-core-components=2.0.0=pypi_0
dash-html-components=2.0.0=pypi_0
dash-table=5.0.0=pypi_0
decorator=5.2.1=pypi_0
diffusers=0.32.2=pypi_0
dnspython=2.6.1=pypi_0
docker-pycreds=0.4.0=pypi_0
einops=0.3.0=pypi_0
email-validator=2.1.1=pypi_0
exceptiongroup=1.2.1=pypi_0
executing=2.2.0=pypi_0
fastapi=0.111.0=pypi_0
fastapi-cli=0.0.4=pypi_0
fastjsonschema=2.21.1=pypi_0
ffmpeg=4.3=hf484d3e_0
ffmpy=0.3.2=pypi_0
filelock=3.15.1=pypi_0
flask=3.0.3=pypi_0
fonttools=4.53.0=pypi_0
freetype=2.12.1=h4a9f257_0
frozenlist=1.4.1=pypi_0
fsspec=2024.6.0=pypi_0
ftfy=6.2.0=pypi_0
future=1.0.0=pypi_0
gitdb=4.0.11=pypi_0
gitpython=3.1.43=pypi_0
gmp=6.2.1=h295c915_3
gnutls=3.6.15=he1e5248_0
google-auth=2.30.0=pypi_0
google-auth-oauthlib=1.0.0=pypi_0
gradio=3.16.2=pypi_0
grpcio=1.64.1=pypi_0
h11=0.14.0=pypi_0
h5py=3.11.0=pypi_0
httpcore=1.0.5=pypi_0
httptools=0.6.1=pypi_0
httpx=0.27.0=pypi_0
huggingface-hub=0.23.3=pypi_0
idna=3.7=py38h06a4308_0
imagecodecs=2023.3.16=pypi_0
imageio=2.9.0=pypi_0
imageio-ffmpeg=0.4.2=pypi_0
importlib-metadata=7.1.0=pypi_0
importlib-resources=6.4.0=pypi_0
intel-openmp=2021.4.0=h06a4308_3561
invisible-watermark=0.2.0=pypi_0
ipython=8.12.3=pypi_0
ipywidgets=8.1.5=pypi_0
itsdangerous=2.2.0=pypi_0
jedi=0.19.2=pypi_0
jinja2=3.1.4=pypi_0
joblib=1.4.2=pypi_0
jpeg=9e=h5eee18b_1
jsonschema=4.22.0=pypi_0
jsonschema-specifications=2023.12.1=pypi_0
jupyter-core=5.7.2=pypi_0
jupyterlab-widgets=3.0.13=pypi_0
kiwisolver=1.4.5=pypi_0
kornia=0.6.0=pypi_0
lame=3.100=h7b6447c_0
lazy-loader=0.4=pypi_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libdeflate=1.17=h5eee18b_1
libffi=3.3=he6710b0_2
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libiconv=1.16=h5eee18b_3
libidn2=2.3.4=h5eee18b_0
libpng=1.6.39=h5eee18b_0
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.19.0=h5eee18b_0
libtiff=4.5.1=h6a678d5_0
libunistring=0.9.10=h27cfd23_0
libwebp-base=1.3.2=h5eee18b_0
linkify-it-py=2.0.3=pypi_0
lmdb=1.4.1=pypi_0
lz4-c=1.9.4=h6a678d5_1
markdown=3.6=pypi_0
markdown-it-py=3.0.0=pypi_0
markupsafe=2.1.5=pypi_0
matplotlib=3.7.5=pypi_0
matplotlib-inline=0.1.7=pypi_0
mdit-py-plugins=0.4.1=pypi_0
mdurl=0.1.2=pypi_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py38h7f8727e_0
mkl_fft=1.3.1=py38hd3c417c_0
mkl_random=1.2.2=py38h51133e4_0
multidict=6.0.5=pypi_0
narwhals=1.28.0=pypi_0
nbformat=5.10.4=pypi_0
ncurses=6.4=h6a678d5_0
nest-asyncio=1.6.0=pypi_0
nettle=3.7.3=hbbd107a_1
networkx=3.1=pypi_0
numpy=1.23.1=py38h6c91a56_0
numpy-base=1.23.1=py38ha15fc14_0
oauthlib=3.2.2=pypi_0
omegaconf=2.1.1=pypi_0
open-clip-torch=2.0.2=pypi_0
open3d=0.19.0=pypi_0
opencv-contrib-python=4.3.0.36=pypi_0
opencv-python=4.10.0.82=pypi_0
opencv-python-headless=4.10.0.82=pypi_0
openh264=2.1.1=h4ff587b_0
openjpeg=2.4.0=h3ad879b_0
openssl=1.1.1w=h7f8727e_0
orjson=3.10.5=pypi_0
packaging=24.1=pypi_0
pandas=2.0.3=pypi_0
parso=0.8.4=pypi_0
pexpect=4.9.0=pypi_0
pickleshare=0.7.5=pypi_0
pillow=10.3.0=py38h5eee18b_0
pip=20.3.3=py38h06a4308_0
pkgutil-resolve-name=1.3.10=pypi_0
platformdirs=4.2.2=pypi_0
plotly=6.0.0=pypi_0
prettytable=3.6.0=pypi_0
prompt-toolkit=3.0.50=pypi_0
protobuf=3.20.3=pypi_0
psutil=6.0.0=pypi_0
ptyprocess=0.7.0=pypi_0
pure-eval=0.2.3=pypi_0
pyarrow=16.1.0=pypi_0
pyasn1=0.6.0=pypi_0
pyasn1-modules=0.4.0=pypi_0
pycryptodome=3.20.0=pypi_0
pydantic=2.7.4=pypi_0
pydantic-core=2.18.4=pypi_0
pydeck=0.9.1=pypi_0
pydeprecate=0.3.1=pypi_0
pydub=0.25.1=pypi_0
pygments=2.18.0=pypi_0
pympler=1.0.1=pypi_0
pyparsing=3.1.2=pypi_0
pyquaternion=0.9.9=pypi_0
pysocks=1.7.1=py38h06a4308_0
python=3.8.5=h7579374_1
python-dateutil=2.9.0.post0=pypi_0
python-dotenv=1.0.1=pypi_0
python-multipart=0.0.9=pypi_0
pytorch=1.12.1=py3.8_cuda11.3_cudnn8.3.2_0
pytorch-lightning=1.5.0=pypi_0
pytorch-mutex=1.0=cuda
pytz=2024.1=pypi_0
pywavelets=1.4.1=pypi_0
pyyaml=6.0.1=pypi_0
qudida=0.0.4=pypi_0
readline=8.2=h5eee18b_0
referencing=0.35.1=pypi_0
regex=2024.5.15=pypi_0
requests=2.32.2=py38h06a4308_0
requests-oauthlib=2.0.0=pypi_0
retrying=1.3.4=pypi_0
rich=13.7.1=pypi_0
rpds-py=0.18.1=pypi_0
rsa=4.9=pypi_0
safetensors=0.5.3=pypi_0
scikit-image=0.18.0=pypi_0
scikit-learn=1.3.2=pypi_0
scipy=1.9.1=pypi_0
semver=3.0.2=pypi_0
sentry-sdk=2.12.0=pypi_0
setproctitle=1.3.3=pypi_0
setuptools=69.5.1=py38h06a4308_0
shellingham=1.5.4=pypi_0
six=1.16.0=pyhd3eb1b0_1
smmap=5.0.1=pypi_0
sniffio=1.3.1=pypi_0
sqlite=3.45.3=h5eee18b_0
stack-data=0.6.3=pypi_0
starlette=0.37.2=pypi_0
streamlit=1.12.1=pypi_0
streamlit-drawable-canvas=0.8.0=pypi_0
tensorboard=2.14.0=pypi_0
tensorboard-data-server=0.7.2=pypi_0
test-tube=0.7.5=pypi_0
threadpoolctl=3.5.0=pypi_0
tifffile=2023.7.10=pypi_0
timm=0.6.12=pypi_0
tk=8.6.14=h39e8969_0
tokenizers=0.12.1=pypi_0
toml=0.10.2=pypi_0
toolz=0.12.1=pypi_0
torchmetrics=0.6.0=pypi_0
torchvision=0.13.1=py38_cu113
tornado=6.4.1=pypi_0
tqdm=4.66.4=pypi_0
traitlets=5.14.3=pypi_0
transformers=4.19.2=pypi_0
typer=0.12.3=pypi_0
typing_extensions=4.11.0=py38h06a4308_0
tzdata=2024.1=pypi_0
tzlocal=5.2=pypi_0
uc-micro-py=1.0.3=pypi_0
ujson=5.10.0=pypi_0
urllib3=2.2.1=py38h06a4308_0
uvicorn=0.30.1=pypi_0
uvloop=0.19.0=pypi_0
validators=0.28.3=pypi_0
wandb=0.17.5=pypi_0
watchdog=4.0.1=pypi_0
watchfiles=0.22.0=pypi_0
wcwidth=0.2.13=pypi_0
webdataset=0.2.5=pypi_0
websockets=12.0=pypi_0
werkzeug=3.0.3=pypi_0
wheel=0.43.0=py38h06a4308_0
widgetsnbextension=4.0.13=pypi_0
xz=5.4.6=h5eee18b_1
yapf=0.32.0=pypi_0
yarl=1.9.4=pypi_0
zipp=3.19.2=pypi_0
zlib=1.2.13=h5eee18b_1
zstd=1.5.5=hc292b87_2
